refactor(engine): extract speculative draft loaders into a mixin (#454)
#3 maintainability: move the six speculative-strategy draft loaders
(classic/dflash/eagle/mtp/pld/self) plus the draft-path resolver, vocab
check, and the dflash attention-version helper (~700 lines) into
engine/speculative_loaders.py as SpeculativeLoaderMixin. ModelManager now
inherits it, so the methods bind self and behave identically. The module
avoids a circular import by never importing model_manager at top level; the
one shared free helper (_load_with_model_type_fallback) is pulled in lazily
at call time. _resolve_attention_causal is re-exported from model_manager
for test back-compat.
